fix: toxic wallets INT4/i64 type mismatch + real postgres tuning
- toxic.rs: local_top_entries/total_entries are INT4 in Postgres but decoded as
  i64 (INT8). Added ::bigint casts to fix the startup warning.
- docker-compose: updated the existing postgres command: tuning for the 31GB
  system (shared_buffers 256MB->4GB, cache 1GB->16GB, work_mem 16MB->64MB,
  maintenance 64MB->256MB).

+++ operator/src/experiment/toxic.rs
@@ -211,7 +211,7 @@ impl ToxicFlowDetector {
         let rows = sqlx::query_as::<_, ToxicWalletRow>(
             r#"
             SELECT wallet_address, selection_roi, post_promotion_roi,
-                   local_top_entries, total_entries, is_toxic,
+                   local_top_entries::bigint, total_entries::bigint, is_toxic,
                    toxic_reason, detected_at
             FROM toxic_wallets
             WHERE is_toxic = TRUE
